+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project uses the changelog in accordance with [keepchangelog](http://keepachangelog.com/). Please use this to write notable changes, which is not the same as git commit log...
 
 ## [unreleased][unreleased]
+- Memory dump color templates (@iceman)
+- Multilanguage support (@iceman)
 - Locked scrolling of dumpfiles
 
 ## [1.0.0.8] - 2018-03-16